ROBOT-1235: Changes related to last review
